+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ##### 2.0 (not yet released)
 
 * an error message is now shown when attempting to create a new map with the same name as already existing one
+* added maps images of all regions and some caves/connectors so that the user doesn't have to download them manually
+    * these will be only copied if "The Long Dark Helpers/Maps Images" folder doesn't exist yet
 
 ##### 1.1 (26/01/2020)